Rename visited-source and collaboration struct.
